pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "braincgt"
version = "0.1.0"
description = "BrainCGT: A Causal Brain Graph Transformer pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "numpy>=1.23",
  "scipy>=1.10",
  "pandas>=2.0",
  "scikit-learn>=1.3",
  "statsmodels>=0.14",
  "nibabel>=5.1",
  "nilearn>=0.10",
  "nipype>=1.8",
  "loguru>=0.7",
  "tqdm>=4.66",
  "torch>=2.1",
  "torch-geometric>=2.4",
  "torch-scatter>=2.1",
  "neuroHarmonize>=2.4"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["braincgt*"]
